Remove some macros in favor of constexpr arrays

+++ check/src/material.cpp
@@ -5,7 +5,7 @@
 
 void material_init (struct material *material)
 {
-    for (int i = 0; i < NR_PLAYERS; i++)
+    for (size_t i = 0; i < NR_PLAYERS; i++)
     {
         material->score[i] = 0;
     }
